Extract CtagsReader from Ctags to support testing
